Modal pour la formulaire de ligne de frais
