+++ src/main/java/dev/controller/GestionFraisController.java
@@ -4,6 +4,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -49,6 +51,25 @@ public class GestionFraisController {
 		 
 		  return resultat;
 	  }
+	  
+		@GetMapping("{UUID}")
+
+		  public List<NotesDeFraisDto> gestionNotesDeFrais(@PathVariable String UUID) {
+			  List<NoteDeFrais> notesFrais =  this.noteDeFraisRepo.findAll();
+			  List<NotesDeFraisDto> resultat = new ArrayList<>();
+			  
+			  for (NoteDeFrais note : notesFrais ) {
+				  if (note.getId() == Integer.parseInt(UUID.replace("UUID=", "")))
+					  resultat.add(new NotesDeFraisDto(note.getId()+"",note.getMission().getId(),note.getMission().getDate_debut(), note.getMission().getDate_fin(), 
+							  note.getMission().getNature().getNom(), note.getMission().getVille_depart(), 
+							  note.getMission().getVille_arrive(), note.getMission().getTransport(), 
+							  note.getPrix()));
+			  }
+			  
+			  return resultat;
+		  }
+		
+		
 	
 	 
 }
